fix (fetcher): synchronous writes to kafka
I broke cluster, up it again and run test and did not get any error. But I forget to create topic. I remember that Produce() have async behaviour and make Produce() of my wrapper with sync behaviour. For the tiny messages it will be a good option. If we need to send largest messages than we should use async methods i think.