feat(atlantserver): Implement interaction with MongoDB
Now we can retrieve product list from MongoDB. Also was changed protocol for start and limit fields - now they are has int64 type, but was int32. I did it because MongoDB use int64 for skip and limit. Also was added condition when limit is zero - it will be set to 100.